+++ ChangeLog.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Declare the minimum search keyword length in the application's settings.
+
+### Changed
+
+- Upgrade to jQuery version 3.5.1
+- Upgrade to Font Awesome version 5.15.1
+- Improve exception handling in data access objects.
+
 ## [0.2.0] - 2020-12-17
 
 ### Added